+++ tailwind.config.js
@@ -4,7 +4,13 @@ module.exports = {
     "./src/**/*.js",
   ],
   theme: {
-    extend: {},
+    extend: {
+      colors: {
+        chisato: 'rgb(173, 64, 74)',
+        takina: 'rgb(49, 57, 74)',
+        kurumi: 'rgb(229, 201, 157)',
+      }
+    },
   },
   plugins: [],
 }